refactor: implement barrel import
