+++ app/(tabs)/_layout.tsx
@@ -5,7 +5,7 @@ import { HapticTab } from '@/components/haptic-tab';
 import { IconSymbol } from '@/components/ui/icon-symbol';
 import { Colors } from '@/constants/theme';
 import { useColorScheme } from '@/hooks/use-color-scheme';
-import { ErrorBoundary } from '@/src/components/common/ErrorBoundary';
+import { ErrorBoundary } from '@/src/components';
 
 export default function TabLayout() {
   const colorScheme = useColorScheme();
@@ -27,19 +27,19 @@ export default function TabLayout() {
         />
         <Tabs.Screen
           name="search"
-        options={{
-          title: 'Search',
-          tabBarIcon: ({ color }) => <IconSymbol size={28} name="magnifyingglass" color={color} />,
-          href: '/search'
-        }}
-      />
-      <Tabs.Screen
-        name="profile"
+          options={{
+            title: 'Search',
+            tabBarIcon: ({ color }) => <IconSymbol size={28} name="magnifyingglass" color={color} />,
+            href: '/search'
+          }}
+        />
+        <Tabs.Screen
+          name="profile"
           options={{
             title: 'Profile',
             tabBarIcon: ({ color }) => <IconSymbol size={28} name="person.fill" color={color} />,
             href: '/profile/123'
-        }}
+          }}
         />
       </Tabs>
     </ErrorBoundary>
