+++ app/search.tsx
@@ -1,13 +1,9 @@
-import { useRouter } from 'expo-router';
-import React, { useEffect } from 'react';
-import { Alert, StyleSheet, Text, View } from 'react-native';
-import { MobileHeader } from '@/src/components/mobile/MobileHeader';
-import { MobileSearch } from '@/src/components/mobile/MobileSearch';
-import { SearchResultItem } from '@/src/components/mobile/SearchResultCard';
+import { CourseCardSkeleton, MobileHeader, MobileSearch, SearchResultItem, Skeleton } from '@/src/components';
 import { sampleCourse } from '@/src/data/sampleCourse';
 import { useAppStore } from '@/src/store';
-import { Skeleton } from '@/src/components/ui/Skeleton';
-import { CourseCardSkeleton } from '@/src/components/mobile/CourseCardSkeleton';
+import { useRouter } from 'expo-router';
+import React, { useEffect } from 'react';
+import { Alert, StyleSheet, View } from 'react-native';
 
 export default function SearchScreen() {
   const router = useRouter();
@@ -17,24 +13,24 @@ export default function SearchScreen() {
     setLoading(true);
 
     const timeoutId = setTimeout(() => {
-        Alert.alert(
-            'Request Timeout',
-            'Loading search results took too long. Please try again.',
-            [
-                { text: 'Retry', onPress: fetchSearchData },
-                { text: 'Cancel', onPress: () => setLoading(false), style: 'cancel' }
-            ]
-        );
+      Alert.alert(
+        'Request Timeout',
+        'Loading search results took too long. Please try again.',
+        [
+          { text: 'Retry', onPress: fetchSearchData },
+          { text: 'Cancel', onPress: () => setLoading(false), style: 'cancel' }
+        ]
+      );
     }, 10000);
 
     const successId = setTimeout(() => {
-        clearTimeout(timeoutId);
-        setLoading(false);
+      clearTimeout(timeoutId);
+      setLoading(false);
     }, 1200);
 
     return () => {
-        clearTimeout(timeoutId);
-        clearTimeout(successId);
+      clearTimeout(timeoutId);
+      clearTimeout(successId);
     };
   };
 
@@ -58,7 +54,7 @@ export default function SearchScreen() {
         <MobileHeader title="Search" showBack />
         <View style={styles.skeletonContainer}>
           <View style={{ flexDirection: 'row', alignItems: 'center', width: '100%', paddingHorizontal: 16, marginBottom: 24, gap: 12 }}>
-             <Skeleton width="100%" height={50} borderRadius={25} />
+            <Skeleton width="100%" height={50} borderRadius={25} />
           </View>
           <CourseCardSkeleton />
           <CourseCardSkeleton />
