+++ src/hooks/index.ts
@@ -1,4 +1,22 @@
+export * from "./useAnalytics";
 export { AuthProvider, useAuth } from "./useAuth";
+export * from "./useBiometricAuth";
+export * from "./useCamera";
+export * from "./useCourseProgress";
+export * from "./useDynamicFontSize";
+export * from "./useGestures";
+export * from "./useHapticFeedback";
+export * from "./useInAppPurchase";
+export * from "./useLongPress";
+export * from "./useMemoryMonitor";
+export * from "./useNetworkStatus";
 export * from "./useNotificationPermission";
+export * from "./useOfflineData";
+export * from "./usePendingRequests";
+export * from "./usePictureInPicture";
+export * from "./usePinchZoom";
+export * from "./useSafeArea";
+export * from "./useScreenReader";
+export * from "./useSwipe";
+export * from "./useVideoGestures";
 export * from "./useVoiceRecognition";
-
