+++ src/screens/index.ts
@@ -1,0 +1,3 @@
+export { MobileLogin } from "../pages/mobile/MobileLogin";
+export { PaymentHistory } from "../pages/mobile/PaymentHistory";
+export { default as SettingsScreen } from "../pages/mobile/Settings";
